+++ packages/yourloops/models/locales.ts
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-export type LanguageCodes = 'en' | 'de' | 'es' | 'fr' | 'it' | 'nl'
+export type LanguageCodes = 'en' | 'de' | 'es' | 'fr' | 'it' | 'nl' | 'jp'
 
 export interface Country {
   code: string